refactor: integrate xxx_syncrepl.go into control.go/response.go for maintainability #447
